+++ src/main/groovy/com/github/andygoossens/gradle/plugins/ModernizerTask.groovy
@@ -21,7 +21,6 @@ import groovy.util.logging.Slf4j
 import org.gaul.modernizer_maven_plugin.Violation
 import org.gradle.api.GradleException
 import org.gradle.api.GradleScriptException
-import org.gradle.api.Project
 import org.gradle.api.file.FileCollection
 import org.gradle.api.tasks.Internal
 import org.xml.sax.SAXException
@@ -37,9 +36,6 @@ class ModernizerTask extends AbstractModernizerTask {
 
     private static final String CLASSPATH_PREFIX = "classpath:"
 
-    @Internal
-    Project project
-    
     @Internal
     ModernizerPluginExtension extension